backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
motor>=3.4.0
pymongo>=4.6.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.0
pydantic>=2.6.0
python-multipart>=0.0.9
openai>=1.20.0
python-dotenv>=1.0.0
httpx>=0.27.0
tenacity>=8.2.0
sentence-transformers>=2.7.0
faiss-cpu>=1.7.4
